move ITrialAppMetadata to folder
