+++ src/explorer/trialApp/ITrialAppMetadata.ts
